Move wallet test into sargon android
