+++ jvm/sargon-android/build.gradle.kts
@@ -50,12 +50,19 @@ cargoNdk {
     librariesNames = arrayListOf("libsargon.so")
 }
 
+tasks.withType<Test> {
+    useJUnitPlatform()
+}
+
 dependencies {
     // Cannot use version catalogues here. For some reason when published to Maven,
     // the jna dependency cannot be resolved
     implementation("net.java.dev.jna:jna:5.13.0@aar")
     implementation("androidx.annotation:annotation:1.7.1")
     implementation("androidx.compose.ui:ui-tooling-preview-android:1.6.2")
+
+    testRuntimeOnly(project(":sargon-desktop-debug"))
+    testImplementation(libs.junit)
 }
 
 publishing {
